+++ CHANGES.md
@@ -1,3 +1,5 @@
+### Version 2.0.5
+
 ### Version 2.0.4
  * Added DuplicateKeyDetector class.
  * ImageListPreference image widget is now right-biased.